ticket 5

+++ models/reviews-models.js
@@ -10,7 +10,7 @@ exports.fetchReviews = (restaurant_id) => {
 }
 
 exports.fetchAllReviews = () => {
-    return connection.query(`SELECT reviews.review_id, title, category, review_img_url, reviews.created_at, reviews.votes, designer, owner,   COUNT(comment_id) AS comment_count FROM reviews
+    return connection.query(`SELECT reviews.review_id, title, category, review_img_url, reviews.created_at, reviews.votes, designer, owner,   COUNT(comment_id)::INT AS comment_count FROM reviews
     LEFT JOIN comments ON comments.review_id = reviews.review_id
     GROUP BY reviews.review_id
     ORDER BY created_at DESC;`).then((result) =>{
